pyproject.toml
[project]
name = "graph-rag-neo4j-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2txt>=0.9",
    "fastapi>=0.122.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-experimental>=0.4.0",
    "langchain-openai>=1.1.0",
    "neo4j>=6.0.3",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]
